+++ drivers/blk/dwmmc-host/src/host2/transaction.rs
@@ -242,7 +242,6 @@ impl sdio_host2::SdioHost for DwMmc {
                     Ok(CommandResponseProgress::Complete(_)) => {
                         Ok(sdio_host2::RequestProgress::WaitingForIrq)
                     }
-                    Ok(_) => Ok(self.pending_transaction_progress()),
                     Err(err) => {
                         self.complete_host2_transaction_request(request);
                         Ok(sdio_host2::RequestProgress::Complete(Err(
@@ -274,7 +273,6 @@ impl sdio_host2::SdioHost for DwMmc {
                     Ok(DataCommandProgress::Complete(_)) => {
                         Ok(sdio_host2::RequestProgress::WaitingForIrq)
                     }
-                    Ok(_) => Ok(self.pending_transaction_progress()),
                     Err(err) => {
                         let _ = self.abort_host2_transaction_request(request);
                         Ok(sdio_host2::RequestProgress::Complete(Err(